BackEnd/requirements.txt
Django>=1.8,<2.0
psycopg2
djangorestframework
markdown
pygments
django-filter
django-cleanup
Pillow
uwsgi